build: update compose.yaml and setup_env.sh

+++ setup_env.sh
@@ -4,12 +4,18 @@
 POSTGRES_USER="bloguser"
 POSTGRES_PASSWORD=$(openssl rand -hex 16)
 POSTGRES_DB="blogdb"
-POSTGRES_HOST_PORT=5432
+POSTGRES_PORT=5432
+
+POSTGRES_HEALTHCHECK_RETRIES=10
+POSTGRES_HEALTHCHECK_INTERVAL="3s"
 
 # pgAdmin service configurations
 PGADMIN_EMAIL="pgadmin@example.com"
 PGADMIN_PASSWORD=$(openssl rand -hex 16)
 PGADMIN_PORT=8080
+
+# Server configurations
+SERVER_PORT=80
 
 #####################################
 
@@ -24,10 +30,17 @@ function check_port_availability() {
 }
 
 # check ports availability
-check_port_availability ${POSTGRES_HOST_PORT}
+check_port_availability ${POSTGRES_PORT}
 check_port_availability ${PGADMIN_PORT}
+check_port_availability ${SERVER_PORT_EXTERNAL}
 
 # Generate the .env configuration file
+if test -e ".env"
+then
+    echo ".env is already exists"
+    exit 1
+fi
+
 cat > .env <<EOF
 # PostgreSQL username
 POSTGRES_USER=${POSTGRES_USER}
@@ -39,7 +52,13 @@ POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
 POSTGRES_DB=${POSTGRES_DB}
 
 # PostgreSQL host port (used for host access to the container)
-POSTGRES_HOST_PORT=${POSTGRES_HOST_PORT}
+POSTGRES_PORT=${POSTGRES_PORT}
+
+# PostgreSQL health check retry times
+POSTGRES_HEALTHCHECK_RETRIES=${POSTGRES_HEALTHCHECK_RETRIES}
+
+# PostgreSQL health check retry interval
+POSTGRES_HEALTHCHECK_INTERVAL=${POSTGRES_HEALTHCHECK_INTERVAL}
 
 # pgAdmin default login email (used to log in to pgAdmin)
 PGADMIN_EMAIL=${PGADMIN_EMAIL}
@@ -49,6 +68,9 @@ PGADMIN_PASSWORD=${PGADMIN_PASSWORD}
 
 # pgAdmin host port (used for host access to the container)
 PGADMIN_PORT=${PGADMIN_PORT}
+
+# Server host port (used for host access to the server)
+SERVER_PORT=${SERVER_PORT}
 EOF
 
 echo "Successfully generated .env"
